requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2
pandas